# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ErrorVo": {
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"messageCode": {
						"type": "string"
					},
					"events": {
						"items": {
							"$ref": "#/components/schemas/EventParamsType"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"messageCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse_any_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"App": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"siteUrl": {
						"type": "string",
						"nullable": true
					},
					"detailDescription": {
						"type": "string",
						"nullable": true
					},
					"bannerUrl": {
						"type": "string",
						"nullable": true
					},
					"promotionalText": {
						"type": "string",
						"nullable": true
					},
					"cardUrl": {
						"type": "string",
						"nullable": true
					},
					"blockChainId": {
						"type": "string"
					},
					"iconUrl": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"appId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"siteUrl",
					"detailDescription",
					"bannerUrl",
					"promotionalText",
					"cardUrl",
					"blockChainId",
					"iconUrl",
					"name",
					"appId"
				],
				"type": "object",
				"description": "Model App\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_App_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/App"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateAppRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "앱 이름",
						"example": "appName",
						"minLength": 1,
						"maxLength": 50
					},
					"iconUrl": {
						"type": "string",
						"example": "https://example.com/example/exampleexample",
						"pattern": "[^(?:/(?:\\*{1,2}|\\w+\\*{0,2}))+/?$]"
					},
					"blockChainId": {
						"type": "string",
						"description": "체인",
						"example": "ETH"
					}
				},
				"required": [
					"name",
					"iconUrl",
					"blockChainId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_CreateAppRequest.blockChainId_": {
				"$ref": "#/components/schemas/Pick_CreateAppRequest.Exclude_keyofCreateAppRequest.blockChainId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateAppRequest": {
				"$ref": "#/components/schemas/Omit_CreateAppRequest.blockChainId_"
			},
...
			"AppDetailResponseType": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"iconUrl": {
						"type": "string",
						"nullable": true
					},
					"blockChainId": {
						"type": "string"
					},
					"cardUrl": {
						"type": "string",
						"nullable": true
					},
					"promotionalText": {
						"type": "string",
						"nullable": true
					},
					"bannerUrl": {
						"type": "string",
						"nullable": true
					},
					"screenUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"detailDescription": {
						"type": "string",
						"nullable": true
					},
					"siteUrl": {
						"type": "string",
						"nullable": true
					},
					"platformUrls": {
						"items": {
							"$ref": "#/components/schemas/PlatformUrl"
						},
						"type": "array"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/ReleaseStatusCodeResponseType",
						"description": "심사상태코드. 코드그룹(APRV_STAT), REQUEST: 심사요청, IN_REVIEW: 심사중, REJECTED: 반려, RELEASED: 출시, CANCELED: 취소됨"
					}
				},
				"required": [
					"appId",
					"name",
					"iconUrl",
					"blockChainId",
					"cardUrl",
					"promotionalText",
					"bannerUrl",
					"screenUrls",
					"detailDescription",
					"siteUrl",
					"createdId",
					"createdAt",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_AppDetailResponseType_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppDetailResponseType"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"FindManyAppsQueryResponse": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id"
					},
					"name": {
						"type": "string",
						"description": "앱 이름"
					},
					"iconUrl": {
						"type": "string",
						"nullable": true,
						"description": "앱 아이콘 url"
					},
					"appWorkspace": {
						"items": {
							"properties": {
								"defaultYn": {
									"type": "string",
									"description": "디폴트 워크스페이스 여부"
								},
								"workspaceId": {
									"type": "number",
									"format": "double",
									"description": "워크스페이스 id"
								}
							},
							"required": [
								"defaultYn",
								"workspaceId"
							],
							"type": "object"
						},
						"type": "array",
						"description": "앱 하위 워크 스페이스 목록"
					},
					"appWallet": {}
				},
				"required": [
					"appId",
					"name",
					"iconUrl",
					"appWorkspace",
					"appWallet"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_FindManyAppsQueryResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FindManyAppsQueryResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"Webhook": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"requireYn": {
						"type": "string"
					},
					"retryCount": {
						"type": "number",
						"format": "double"
					},
					"webhookKindCd": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"webhookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"requireYn",
					"retryCount",
					"webhookKindCd",
					"name",
					"webhookId"
				],
				"type": "object",
				"description": "Model Webhook\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_Webhook_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/Webhook"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateWebhookRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "웹훅명"
					},
					"webhookKindCode": {
						"type": "string",
						"description": "웹훅 분류 코드\r\n코드그룹 : APS_WHK\r\n코드 : SWAP(게임재화 스왑) , TOKEN_IN(토큰 입금) , NFT_MINT(NFT 생성) , REVIEW_RESULT(심사결과 전송)"
					},
					"requireYn": {
						"type": "string",
						"description": "필수 여부 Y/N"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "로그인한 회원 id"
					}
				},
				"required": [
					"name",
					"webhookKindCode",
					"requireYn",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_Webhook-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Webhook"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"AppWebhook": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"encryptionYn": {
						"type": "string"
					},
					"endpointUrl": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"appId": {
						"type": "number",
						"format": "double"
					},
					"webhookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"encryptionYn",
					"endpointUrl",
					"name",
					"profile",
					"appId",
					"webhookId"
				],
				"type": "object",
				"description": "Model AppWebhook\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_AppWebhook_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppWebhook"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateAppWebhookRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱id"
					},
					"profile": {
						"type": "string",
						"description": "프로필 : DEV(개발환경,테스트) , PRD(운영)"
					},
					"name": {
						"type": "string",
						"description": "웹훅명",
						"minLength": 1,
						"maxLength": 20
					},
					"encryptionYn": {
						"type": "string",
						"description": "구간암호화 여부 Y/N"
					},
					"endpointUrl": {
						"type": "string",
						"description": "엔드포인트 URL"
					},
					"webhookId": {
						"type": "number",
						"format": "double",
						"description": "웹훅 id"
					}
				},
				"required": [
					"appId",
					"profile",
					"name",
					"encryptionYn",
					"endpointUrl",
					"webhookId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAppWebhookRequest": {
				"$ref": "#/components/schemas/CreateAppWebhookRequest"
			},
			"CommonResponse_AppWebhook-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/AppWebhook"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"DeleteAppWebhookRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double"
					},
					"profile": {
						"type": "string",
						"enum": [
							"DEV",
							"PRD"
						]
					},
					"webhookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"appId",
					"profile",
					"webhookId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse_AppWebhookLogsQueryResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/AppWebhookLogsQueryResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_AppWebhookLog_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppWebhookLog"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"ResendAppWebhookRequest": {
				"properties": {
					"no": {
						"type": "number",
						"format": "double"
					},
					"processDt": {
						"type": "string"
					}
				},
				"required": [
					"no",
					"processDt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"devAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			},
			"token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"cookie": {
				"type": "apiKey",
				"name": "accessToken",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "platpome-console-backend",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/apps": {
			"post": {
				"operationId": "CreateApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "콘솔 메인 홈 > 앱 등록 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindManyApps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindManyAppsQueryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "콘솔 회원이 속한 앱 목록 조회 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "앱 이름 검색 값 (포함 검색)",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/apps/{appId}": {
			"put": {
				"operationId": "UpdateApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "앱 정보 수정 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAppRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppDetailResponseType_"
								}
							}
						}
					}
				},
				"description": "앱 상세 조회 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "= 앱 아이디",
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/webhooks": {
			"post": {
				"operationId": "CreateWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_Webhook_"
								}
							}
						}
					}
				},
				"description": "** 웹훅 등록은 시스템으로 처리 **\r\n웹훅 등록 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"deprecated": true,
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWebhookRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindManyWebhooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_Webhook-Array_"
								}
							}
						}
					}
				},
				"description": "웹훅 목록 조회 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/webhooks/apps": {
			"post": {
				"operationId": "CreateAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 등록 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppWebhookRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 수정 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAppWebhookRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAppWebhookRequest"
							}
						}
					}
				}
			}
		},
		"/webhooks/apps/{appId}": {
			"get": {
				"operationId": "FindManyAppWebhooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook-Array_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 목록 조회 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "'DEV' | 'PRD'",
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"DEV",
								"PRD"
							]
						}
					}
				]
			}
		},
		"/webhooks/logs/{appId}": {
			"get": {
				"operationId": "FindManyAppWebhookLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhookLogsQueryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "웹훅 로그 목록 조회 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "앱 id",
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "YYYY-MM-DD",
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "YYYY-MM-DD",
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "'DEV' | 'PRD'",
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"DEV",
								"PRD"
							]
						}
					},
					{
						"description": "'SWAP' | 'TOKEN_IN' | 'NFT_MINT' | 'REVIEW_RESULT'",
						"in": "query",
						"name": "webhookKindCd",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"SWAP",
								"TOKEN_IN",
								"NFT_MINT",
								"REVIEW_RESULT"
							]
						}
					},
					{
						"description": "'Success'|'Fail'",
						"in": "query",
						"name": "logStatus",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"Success",
								"Fail"
							]
						}
					}
				]
			}
		},
		"/webhooks/logs/{appId}/{no}": {
			"get": {
				"operationId": "FindAppWebhookLog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhookLog_"
								}
							}
						}
					}
				},
				"description": "웹훅 로그 상세 조회 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "no",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/webhooks/resend": {
			"put": {
				"operationId": "ResendAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_any_"
								}
							}
						}
					}
				},
				"description": "웹훅 재전송",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResendAppWebhookRequest"
							}
						}
					}
				}
			}
		},
		"/webhooks/excel/logs/{appId}": {
			"get": {
				"operationId": "DownloadAppWebhookLogsExcel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"error": {
											"$ref": "#/components/schemas/ErrorVo"
										},
										"data": {}
									},
									"required": [
										"total",
										"error",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "웹 훅 목록 엑셀 다운로드 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "YYYY-MM-DD",
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "YYYY-MM-DD",
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "'DEV' | 'PRD'",
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"DEV",
								"PRD"
							]
						}
					},
					{
						"description": "'SWAP' | 'TOKEN_IN' | 'NFT_MINT' | 'REVIEW_RESULT'",
						"in": "query",
						"name": "webhookKindCd",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"SWAP",
								"TOKEN_IN",
								"NFT_MINT",
								"REVIEW_RESULT"
							]
						}
					},
					{
						"description": "'Success'|'Fail'",
						"in": "query",
						"name": "logStatus",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"Success",
								"Fail"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}